# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Engage API",
    "version": "1.0.0",
    "description": "選挙活動管理プラットフォーム Engage のAPIドキュメント"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/api/admin/organizations": {
      "get": {
        "summary": "組織一覧取得",
        "responses": {
          "200": {
            "description": "組織一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "組織更新",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新された組織",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "組織削除",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "削除された組織",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/participant-lead-management": {
      "get": {
        "summary": "紐付け候補一覧取得",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventId": {
                    "type": "string"
                  },
                  "page": {
                    "type": "integer"
                  },
                  "limit": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "紐付け候補一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "participations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Participation"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "totalCount": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "紐付け候補の自動生成",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "participationIds": {
                    "type": "array"
                  },
                  "algorithm": {
                    "type": "string"
                  },
                  "confidenceThreshold": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "マッチング分析結果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "participationId": {
                            "type": "string"
                          },
                          "candidatesFound": {
                            "type": "integer"
                          },
                          "candidates": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/MatchCandidate"
                            }
                          }
                        }
                      }
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "processedParticipations": {
                          "type": "integer"
                        },
                        "totalCandidatesFound": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Organization": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Participation": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "eventId": { "type": "string" },
          "leadId": { "type": "string" },
          "participantName": { "type": "string" },
          "participantEmail": { "type": "string" },
          "status": { "type": "string" },
          "registeredAt": { "type": "string", "format": "date-time" }
        }
      },
      "MatchCandidate": {
        "type": "object",
        "properties": {
          "leadId": { "type": "string" },
          "confidence": { "type": "number" },
          "matchedFields": { "type": "object" },
          "leadName": { "type": "string" },
          "leadEmail": { "type": "string" }
        }
      },
...
    }
  },
...
}